Refactor to getmap
