+++ example-app/src/androidTest/java/com/example/getmap/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.example_app
+package com.example.getmap
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
