+++ example-app/src/main/java/com/example/getmap/MultiPolygonDto.kt
@@ -1,3 +1,3 @@
-package com.example.example_app
+package com.example.getmap
 
 class MultiPolygonDto(var type: String, var coordinates: List<List<List<List<Double>>>>)
