+++ example-app/src/main/java/com/example/getmap/PolyObject.kt
@@ -1,4 +1,4 @@
-package com.example.example_app
+package com.example.getmap
 
 import java.math.BigDecimal
 import java.time.OffsetDateTime
