+++ example-app/src/main/java/com/example/getmap/PolygonDTO.kt
@@ -1,3 +1,3 @@
-package com.example.example_app
+package com.example.getmap
 
 class PolygonDTO(var type: String, var coordinates: List<List<List<Double>>>)
