+++ example-app/src/main/java/com/example/getmap/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example_app
+package com.example.getmap
 
 import GetApp.Client.models.MapConfigDto
 import PasswordDialog
@@ -11,7 +11,6 @@ import android.util.Log
 import android.view.MotionEvent
 import android.view.View
 import android.view.animation.LinearInterpolator
-import android.view.animation.RotateAnimation
 import android.view.inputmethod.InputMethodManager
 import android.widget.Button
 import android.widget.ImageButton
@@ -21,13 +20,12 @@ import android.widget.Toast
 import android.widget.ToggleButton
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.graphics.rotationMatrix
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.example_app.matomo.MatomoTracker
-import com.example.example_app.models.ConfigParam.NebulaParamAdapter
-import com.example.example_app.models.ConfigParam.NebulaParam
+import com.example.getmap.matomo.MatomoTracker
+import com.example.getmap.models.ConfigParam.NebulaParamAdapter
+import com.example.getmap.models.ConfigParam.NebulaParam
 import com.ngsoft.getapp.sdk.Configuration
 import com.ngsoft.getapp.sdk.GetMapService
 import kotlinx.coroutines.Dispatchers
