+++ example-app/src/main/java/com/example/getmap/matomo/MatomoTracker.kt
@@ -1,4 +1,4 @@
-package com.example.example_app.matomo
+package com.example.getmap.matomo
 
 import android.content.ComponentCallbacks2
 import android.content.Context
