+++ example-app/src/main/java/com/example/getmap/models/ConfigParam.kt
@@ -1,25 +1,22 @@
-package com.example.example_app.models
+package com.example.getmap.models
 
 import android.app.AlertDialog
 import android.content.Context
 import android.text.Editable
-import android.text.InputType
 import android.text.InputType.TYPE_CLASS_TEXT
 import android.text.InputType.TYPE_CLASS_NUMBER
 import android.text.InputType.TYPE_TEXT_VARIATION_PASSWORD
 import android.text.TextWatcher
 import android.text.method.HideReturnsTransformationMethod
 import android.text.method.PasswordTransformationMethod
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageButton
-import android.widget.Spinner
 import android.widget.TextView
 import androidx.cardview.widget.CardView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.example_app.R
+import com.example.getmap.R
 import com.google.android.material.textfield.TextInputEditText
 
 class ConfigParam {
