+++ example-app/src/main/java/com/example/getmap/models/MapDataMetaData.kt
